Consolidate articles under /articles/ path with SEO redirects and update author bio
- Move all 6 articles to /articles/ with proper canonical URLs and JSON-LD schema updates
- Move case-studies collection to /articles/case-studies/ with matching article structure
- Create 301 redirect stubs at old paths for SEO authority transfer
- Update all navigation links across header, footer, articles, and sitemap
- Update case study pages with consistent article navigation (back-to-articles, RulesApplied, AuthorBio)
- Replace author bio with comprehensive professional background (20+ years, technologies, expertise)
- Apply long author bio variant to all article pages for consistency
- Build verified: 0 errors, 92 HTML files, 5382 words indexed
- All old /case-studies paths redirect with permanent-redirect metadata

+++ site/app/articles/what-is-diff-based-analysis/page.tsx
@@ -628,11 +628,12 @@ export default function WhatIsDiffBasedAnalysisPage() {
           </div>
 
           <RulesApplied ids={["GCI0001", "GCI0003", "GCI0004"]} />
-          <AuthorBio variant="short" />
+          <AuthorBio variant="long" />
         </div>
       </main>
       <Footer />
     </>
   );
 }
+
